enhancement: only include tests that end with .test.ts
this pr changes how test files are detected and included in the build artifacts.
The purpose of this is to be able to abstract common shared test logic away from test files
without making those micro-libraries part of the artifacts.

It also moves the tests to a folder and renames the script to README.md as content entry point.

beside the test file consideration this pr renames the client baseUrl to platformUrl.